+++ example/index.php
@@ -8,6 +8,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
+use JoliNotif\Driver\AppleScriptDriver;
 use JoliNotif\Driver\GrowlNotifyDriver;
 use JoliNotif\Driver\NotifySendDriver;
 use JoliNotif\Driver\TerminalNotifierDriver;
@@ -17,6 +18,7 @@ use JoliNotif\Notifier;
 require __DIR__.'/../vendor/autoload.php';
 
 $notifier = new Notifier([
+    new AppleScriptDriver(),
     new TerminalNotifierDriver(),
     new GrowlNotifyDriver(),
     new NotifySendDriver(),